requirements.txt
fastapi~=0.115.11
uvicorn
sqlalchemy~=2.0.39
psycopg2-binary
pydantic~=2.10.6
python-jose~=3.4.0
passlib~=1.7.4
pydantic[email]
pydantic-settings~=2.0.3
python-multipart
starlette~=0.46.1
slowapi~=0.1.9
cloudinary~=1.43.0
python-dotenv~=1.0.1
bcrypt
sphinx~=7.2.6
sphinx-rtd-theme~=2.0.0
sphinxcontrib.httpdomain~=1.8.1
pytest~=8.0.2
pytest-cov~=4.1.0
pytest-asyncio~=0.23.5
httpx~=0.27.0
redis~=5.0.1
email-validator~=2.1.0.post1